Fix S-04: split MovieMutatorBase.swift god file
The 972-line MovieMutatorBase.swift mixed Notification.Name extension,
AVMutableMovie extension, helper types (boxSize, dimensionsType,
RefOrSelfCont), PresentationInfo, and MovieMutatorBase logic. Split
into focused files for readability and future refactoring.

- MovieMutatorBase.swift: 972 -> 322 lines (67% reduction)
- New MovieMutatorBase+PresentationInfo.swift: PresentationInfo and
  presentation/time-mapping helpers
- New MovieMutatorBase+Progress.swift: async progressStream()
- New MovieMutatorBase+FormatDescriptions.swift: mediaDimensions*()
- New MovieMutatorBase+Formatting.swift: shortTimeString(), ts(), etc.
- New MovieMutatorTypes.swift: boxSize, dimensionsType, RefOrSelfCont
- New AVMutableMovie+Extensions.swift: range, movHeader, findReferenceURLs
- New Notifications.swift: Notification.Name.movieWasMutated

API signatures and access levels unchanged. Verified with build,
analyze, and 43 MovieMutatorTests/ModelTests (all pass).

Plan: /_plans/S-04_moviemutatorbase_split_plan.md (v4.1)
