+++ cutter2/Models/Notifications.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+extension Notification.Name {
+    static let movieWasMutated = Notification.Name("movieWasMutated")
+}
